Add fix for #252
Added a Try around the awaitResponse method to prevent an Exception
killing the Ammonite Repl. The awaitResponse method throws an Exception
if a match can't be found for a completion.

The Failure case simply returns an empty List[pressy.Member]. This seems
like the simplest and least invasive fix for issue #252, although it may
not be ideal.