Add regression test around #252 and fix formatting
Add fix for Pressy.prefix error in 2.10.x tests

Pressy.prefix throws a java.lang.UnsupportedOperationException when
we call .end on a Position that is not a range in scala 2.10.x..
This is not the case with scala 2.11.x where we can call .end without
any issues.

I have added a check around Position to verify that is a range,
and only then call .end on it.