fix crash when autocompleting an unknown import
When trying to autocomplete an unknown or mistyped import such as:

```
import scala.collction.<tab>
```

Ammonite exits with the following error:

```
@ import scala.collction.Exception in thread "main"
java.lang.ClassCastException: scala.reflect.internal.Trees$Select
    cannot be cast to scala.reflect.internal.Trees$Ident
    at ammonite.repl.interp.Pressy$Run.prefixed(Pressy.scala:135)
    at ammonite.repl.interp.Pressy$$anon$3.complete(Pressy.scala:231)
```

The error is due to casting a scala.reflect.internal.Trees$Select to a
scala.reflect.internal.Trees$Ident. The fix simply tests for the
Trees$Ident type before casting to it.

If the expr type of the Import is not a Trees$Ident then we return no
matches.

Also added a couple of tests for the fix to repl/AutocompleteTests.