Use special base class loader if one is found
To run an Ammonite session with class loader isolation enabled, do

  $ sbt publishLocal
  $ test -e coursier || ( curl -L -o coursier https://git.io/vK2my && chmod +x coursier )
  $ coursier launch \
      -i ammonite \
      -I ammonite:com.lihaoyi:ammonite-api_2.11.8:0.6.3-SNAPSHOT \
      -I ammonite:com.lihaoyi:ammonite-tools_2.11.8:0.6.3-SNAPSHOT \
      com.lihaoyi:ammonite_2.11.8:0.6.3-SNAPSHOT

The -i / -I options passed to coursier make it launch ammonite with a
two-ClassLoader setup. A first loader with ammonite-api /
ammonite-tools in it, and a second one with ammonite, having the former
loader as parent. The reflective call to `getIsolatedTargets` in this PR
allows to find the first loader, and use it as a base loader in the
ammonite session.