Move the unwrapping of top-level `{}` blocks into the REPL-specific code, rather than having it apply to both REPL and scripts.
This was introduced as a REPL feature (to allow people to enter multiple statements
in one REPL block) and is not useful in scripts. Furthermore, it causes problems
such as https://github.com/lihaoyi/Ammonite/issues/453 due to the way we're parsing
out the `{}`s. Those problems do not exist in the REPL since they only arise when
there are subsequent expressions/statements following the unwrapped `{}` code,
which cannot happen in the REPL since after completing a top-level `{}` entering
a newline will cause the block to be submitted immediately, preventing you from
entering further statements