Use Console for SSHD stream redirection (#513)
Previously, logging into an SSH session redirected System.out to the
console. This caused problems with system behaviour, along with not
being safe for multiple sessions.

While Scala's Console object only redirects `println` and not
`System.out.println`, it is thread-local, and so behaves in a more
sensible way when SSH sessions are active.