Properly parse ammonite arguments before or after a scala script
As documented in #618, it was not possible to specify ammonite arguments
if also specifying a scala script. With the change, the following
invocations work:

    amm --silent my-script.sc scriptarg1 scriptarg2
    amm --silent my-script.sc -- scriptarg1 scriptarg2
    amm my-script.sc --silent -- scriptarg1 scriptarg2
    amm my-script.sc --silent scriptarg1 scriptarg2

Additionally, we add assertions to prevent scriptargs from being
specified both before and after the delimiter. The following is
rejected:

    amm my-script.sc --silent scriptarg1 -- scriptarg2

We also add a proper `--help` flag.

Fixes #618