Add Java serialization test (fails with object wrapping)
This adds a test, serializing a closure (`closure`), that itself
references a value computed in a previous command (`a`).

This is similar to what happens when [mapping a function on an RDD](https://spark.apache.org/docs/2.2.0/api/scala/index.html#org.apache.spark.rdd.RDD@map[U](f:T=%3EU)(implicitevidence$3:scala.reflect.ClassTag[U]):org.apache.spark.rdd.RDD[U])
or using a [UDF](https://spark.apache.org/docs/2.2.0/api/scala/index.html#org.apache.spark.sql.expressions.UserDefinedFunction) in Spark: a closure gets
serialized this way, and is sent to a bunch of "executors" that each
deserialize it.

This kind of closure can reference the results of heavy computations. We
want these results to be serialized with the closure, rather than re-computed
on each machine after deserialization.

In the test, `a` acts as the result of the heavy computation. We do not want
the calculation of `a` to re-run when deserializing `closure`.

After creating the closure, the test serializes it, then deserializes it
via a classloader that can load the exact same bytecode as the classloader of
the current session. We call the deserialized closure by reflection, and check
that the heavy computation was not run again.