Add raw code as resource
This commit modifies `AmmonitePlugin`, so that it passes the raw string representation of the tree of the user code to Ammonite. These strings
look like
```scala
List(def <init>(): cmd0.this.Helper = {
  Helper.super.<init>();
  ()
}, private[this] val n: Int = 2, <stable> <accessor> def n: Int = Helper.this.n, def $main(): Iterator[String] = ammonite.repl.ReplBridge.value.Internal.combinePrints(ammonite.repl.ReplBridge.value.Internal.print[Int](Helper.this.n, "n", scala.None)((pprint.TPrint.lambda[Int](((cfg$macro$2: pprint.TPrintColors) => "".+(cfg$macro$2.typeColor.apply(fansi.this.Str.implicitApply("Int")).render).+(""))): pprint.TPrint[Int]), ammonite.repl.ReplBridge.value.tprintColorsImplicit, (ClassTag.Int: scala.reflect.ClassTag[Int]))), override def toString: String = "cmd0")
```

It is then written in a resource file (`cmd1-tree.txt` if the wrapper name is `cmd1`). This allows wrappers to process this tree at runtime. This is done
for class wrapping in later commits. E.g. checking for substrings such as `cmd1.this.cmd0` in this string says whether `cmd1` uses things
from `cmd0`. This is then used to not keep references to unused previous commands, so that these don't hinder serialization of the current one.