Add back object wrapping for… objects
If the user code solely consists of a singleton, like
```scala
object Foo {
  val n = 2
}
```
then it is wrapped in an object rather than a class.

This allows users to wrap some code in objects rather than classes if needed.

In particular, this is useful for macro code (than can't be defined in classes that we instantiate ourselves), or for definitions that are later processed by macros (definitions from classes are harder to handle than those in objects)