Fix JLine dependencies (#851)
jline-$VERSION.jar is a bundle that contains the content of
jline-terminal-$VERSION.jar and other modules. jline-terminal-jna
depends on jline-terminal. This means that the Ammonite build depends on
both jline and jline-terminal and classfiles may be read from either
jar (dependencies on the classpath are usually not in a stable order).
They should be identical so it's not a big problem, but it breaks Zinc
incremental compilation. This commit fixes this by not using the jline
bundle jar anymore.