Injecting import hooks
Rather than having the import hooks hardcoded, this allows to inject
them as a `Map[Seq[String], ImportHook]` from the main.

The aim at this is for ammonite's downstreams to have a mean of
changing the behaviour of the import hooks. In particular, this should
allow mill to inject its own version into the resolution logic, in order
to avoid discrepencies of versions between mill and the loaded
contrib-module