Add a talk from scalawave 2018
Hey, last year I gave a talk about why bash is a bad solution and how to replace it with Ammonite. As there is already a section with the talks I thought we may want to have it there as well. And its quite recent, so in some way it shows that project is alive enough for people to talk about it.