Make InterpAPI an abstract Java class
This changes InterpAPI from a Scala class to a Java class. For this to
work, some of the methods of InterpAPI get more Java-compatible
signatures. In particular, mutable collections become java.util.List.

In order to recover the ease of use of the Scala signatures and
collections, we add extension methods, defined in
InterpBridge.InterpAPIExtensions, so that most former Scala methods can
still be used with InterpAPI.