Don't try to de-duplicate classpath / initial classpath
When creating a compiler instance, Ammonite passes it two class paths:
the "initial" one, and the main one. The former is subject to
whitelisting when --thin is passed, not the latter.

Later on, when adding --scala-version, there can be JARs used internally
by Ammonite that are in both paths. Their classes need to be blacklisted,
but these JARs should still be loaded via the main class path, as any
other JAR added by users.

Before this commit, we were naively removing from the main class path
any JAR already in the initial class path, which was a problem for the
JARs in both paths mentioned above.