Java-ize ScriptInvoker stuff
This moves to interp-api-full some of the logic interfacing with
mainargs to run scripts. interp-api-full is loaded on the fly in the
user session when --scala X.Y.Z is used.

In order to deal with possibly different Scala versions between Ammonite
and userspace, we get the mainargs.ParserForMethods[Any] generated by
the mainargs macro as an Object. We then get an instance of a
ammonite.interp.api.ScriptInvoker by reflection from interp-api-full,
and pass it the ParserForMethods, that it casts back to its proper type.