Java-ize internal Evaluator stuff
Ammonite used to get an Iterator[String] out of the printing logic it
generated. As the Scala version from Ammonite and userspace may now
differ, there's no more shared Scala Iterator class, so we resort to
wrapping the generated Scala Iterator in a Java Iterator in the
generated code, and convert it back to a Scala iterator from Ammonite.