Add back TestBridge
TestBridge lives in "user space" in some tests. It needs to be compiled
with the user space Scala version, rather than the Ammonite Scala
version.

It's put in a dedicated module, and its class path is passed along the
one of amm.compiler and amm.repl.api.full to the cross tests.