Disambiguate compiler-interface module name (#1215)
It's compiled only in Scala 2, and used via dotty compat in Scala 3.
For Scala 2, it depends on Scala 2 pprint. For Scala 3, it depends on
Scala 3 pprint.

Without these changes, only the latter module ends up being published,
which makes ammonite for 2.13.6 pull pprint for Scala 3 (along with
pprint for Scala 2), which is a problem.

Without these changes, the same module
(com.lihaoyi:ammonite-compiler-interface_2.13.6) was built and published
twice. A few commits ago, these were published in the same batch, so
only one (apparently the latter) ended up actually published. Since two
commits on master, these are published in separate batches, which makes
one of them fail (because it tries to publish a module already published
by the other batch).