Fix Scala 3 dependencies to be propagated via moduleDeps
As stated [here](https://github.com/com-lihaoyi/mill/issues/860#issuecomment-974654679)
Mill doesn't handle correctly the Scala 2.13 modules depending on Scala
3 modules. This happens because dependencies are resolved at using
the suffix from the leaf module instead of every module resolving its dependencies
on its own. So if a module compiled with Scala 3 depends on libraries
with the `_3` suffix, once another module using Scala 2.13 depends on it
all the dependencies get transformed to use the `_2.13` suffix.
Sbt maintains instead the original suffixes.
This commit introduces a preprocessing on the dependencies to convert
them to manually add the `_3` suffix