Split up Main.scala into Main.scala/MainRunner.scala/AmmoniteMain.scala (#1229)
This should help make things easier to find.

The dependency order is `AmmoniteMain.scala` depends on `MainRunner.scala` depends on `Main.scala`

The new main method is in `ammonite.AmmoniteMain`. This was chosen to help improve the UX of tools like `jps` which only list the classname and not the package path as part of its output. Previously it would just show a mysterious `Main` JVM process running, with this change it'll show `AmmoniteMain` which is a lot more useful.

A few minor helper functions needed to be moved around to preserve acyclicity of the three files, I moved them all to `ammonite.util`