Fix handling of imports involving synthetic case class objects (#1231)
Fixes #474

Turns out that even though the companion object for a `case class Foo` is marked as synthetic, it can still be imported and used. Thus when determining imports, we should not filter out synthetic symbols 

Added tests including the original repro, as well as a minimized repro. Both tests fail on master, pass on this PR