Use full file path for parse error messages and compiler metadata (#1233)
Fixes https://github.com/com-lihaoyi/Ammonite/issues/646
Fixes https://github.com/com-lihaoyi/Ammonite/issues/493

Previously we constructed a `new io.VirtualFile` with an identical name and path, both short, to ensure stack traces included only the file name. The correct thing to do is to provide separate names and paths, so stack traces can use the file name while compiler errors can use the full file path. This is done differently in Scala2 and Scala3 due to differences in the compiler APIs, but the end result is the same

This makes the parse/compile errors line up more with how the normal Scala compiler reports them, and fixes `sourcecode.File` to properly report the full path of the script file.

Updated the existing LineNumberTests to verify the fixed behavior, and added a new `sourceCodeMetadata.sc` test case to verify the fix applies to the `sourcecode.*` macros. For now `sourcecode.*` macros do not work in the REPL or scripts in Scala3, but that is an unrelated problem we can deal with later

Review by @alexarchambault  and @anatoliykmetyuk