Fix stack traces line numbers in Scala 3
In Scala 2, a compiler plugin updates line numbers in the positions
generated by the compiler. We don't have such a plugin in Scala 3 (and
can't have one it seems), so we post-process the compiler output and
generated artifacts so that they have correct positions. A previous
commit post-processes diagnostics. This one post-processes byte code, so
that line numbers in stack traces are correct.