Update upickle to 3.0.0-M2 (#1320)
This is a binary incompatible change! The next release version after
merge is probably `3.0.0`.

I dropped support for Scala 3.0.x due to the fact, that upickle_3
3.0.0-M2 is build against Scala 3.1.

Ammonite depends also on geny but never made that dependency explicit. I
changed that and added a direct dependency where appropriate. With the
bump of upickle, we implicitly also bump geny from 0.7.0 to 1.0.0. As a
consequence, I also had to update os-lib from 0.8.0 to 0.9.0 as it is
also depending on geny, and only version 0.9.0 is compatible to geny
1.0.0.

Pull request: https://github.com/com-lihaoyi/Ammonite/pull/1320