Fix publishDocs, publishExecutable and skipBloop (#1384)
publishExecutable was using the binary Scala 3 version, which needs to
be the lowest minor version, currently 3.2 It needs instead use the
latest Scala 3 version.
So this introduces a new list (assemblyCrossScalaVersions) which
contains the last Scala 3 version instead.
The same is used on some modules to skipBloop, since the last Scala 3
version depends on crossFullScalaVersion